Update DecimoOitavoExercicio.java
Alterando o tipo das variáveis para double

+++ curso_youtube/src/exercicios/curso/DecimoOitavoExercicio.java
@@ -10,12 +10,12 @@ public class DecimoOitavoExercicio {
 		
 		System.out.print("Digite o tamanho do arquivo: ");
 		
-		int arquivoDownload = scan.nextInt();
-		int conversaoArquivo = arquivoDownload * 8;
+		double arquivoDownload = scan.nextInt();
+		double conversaoArquivo = arquivoDownload * 8;
 		
 		System.out.print("Digite a velocidade da Internet: ");
-		int linkInternet = scan.nextInt();
-		int tempoDownload = conversaoArquivo / linkInternet;
+		double linkInternet = scan.nextInt();
+		double tempoDownload = conversaoArquivo / linkInternet;
 		
 		System.out.print("Esse download tem o prazo estimado de: " + tempoDownload / 60);
